requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
djangorestframework==3.14.0
et-xmlfile==1.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
openpyxl==3.1.1
psycopg2==2.9.5
pydantic==1.10.6
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.15
yarl==1.8.2
